API change: Budget/Noise calculations return BudgetTrace object
This new object holds the PSD and frequency arrays, as well as the
list of all budget sub-noises, accessible via a dictionary interface
and attribute access.

All interfaces and tests are updated to process the new BudgetTrace
object.

A new HDF5 schemata (version 2) is also introduced that stores trace
data in a way that mirrors the BudgetTrace.  The io functions
(save_hdf5/load_hdf5) are updated to write/read this format.  The
load_hdf5 functions continues to support the version 1 SCHEMA.

Closes #59