Residual gas sub-budgets and updates
Adds residual gas damping and additional molecular species to the residual gas calculations. The gas scattering and gas damping are now tracked separately for each species in an excess gas sub-budget. The damping calculation is done in the infinite volume limit by default but adds corrections for squeezed film damping if specified in the yaml file.

H2, N2, H2O, and O2 are considered. The numbers for aLIGO are taken from the link in this commit. The numbers for Aplus and Voyager are copied from aLIGO.  For the existing LIGO facilities the O2 numbers are unknown but are suspected to be low, and are therefore set 10x below the expected noise floor.

Closes #63.